+++ src/main/java/com/example/demo/common/exception/MemberException.java
@@ -1,0 +1,4 @@
+package com.example.demo.common.exception;
+
+public class MemberException {
+}
