+++ src/main/java/com/example/demo/common/exception/code/MemberErrorCode.java
@@ -1,0 +1,4 @@
+package com.example.demo.common.exception.code;
+
+public enum MemberErrorCode {
+}
