+++ src/main/java/com/example/demo/member/service/memberService.java
@@ -1,0 +1,4 @@
+package com.example.demo.member.service;
+
+public class memberService {
+}
