+++ src/main/java/com/example/demo/member/web/memberController.java
@@ -1,0 +1,4 @@
+package com.example.demo.member.web;
+
+public class memberController {
+}
